This commit makes the pulse length modular per protocol
Almost all devices that we have seen have a 0 length of about 294ms.
However, impuls and select-remote use a pulse length of about 146ms.
This means that these latter brands use a pulse length that is
half of the regular length. This commit adds the options to define
a divider for the standard pulse length. All protocols have a
"plslen" of 1, however the impuls protocol will use a "plslen" of
2. This means that the value of 294/2 is used. This commit also
changes the default pulse length from 295ms to 294ms to allow for a
non-decimal devision.