This commit adds several small (internal) features
1) When pilight daemon recieves a repeated code it will communicate
for each received code which number the specific code is in the
total number of repeats.

2) The config file won't be updated anymore while the daemon is
running. When the daemon is started, the config file will be writen
to once and once again on shutdown. All changes are kept inside
the memory. This reduces unnecessary IO.

3) Then pilight daemon was started as a non-root user it showed
a segmentation fault. The daemon will now check if it is started as
a root priveliged user or informs otherwise.

4) A small improvement was added in the options library in which
multiple error return codes are given specific to each error.

5) The pilight logs will now contain microseconds to enable
specific benchmarking of code.