This commit adds multithreaded sending and broadcasting
One drawback of pilight until now was that sending bulk codes
wasn't possible, because sending and broadcasting where blocking
functions. This commit adds two thread safe queues for both
sending and broadcasting. So when a new code is sent, it is added
to the send queue. When an another new code is sent while the
previous one is still busy, it will be added to the queue. This
way, you can control multiple devices way faster. After a message
is send or received, the message to be broadcasted will be added
to the broadcast queue. This means that updating the config and
GUIs will also be handled asynchronously from sending and
receiving. This allows for far faster handling of the core
functions.

Because of this focus on speed and stability, the process-file
capability was dropped. Instead of using process file pilight
now includes basic receivers in perl, python and php so you
can easily build a new process file like script but then also
independent from the main daemon. Would anything fail in the
process file, the daemon won't suffer from it.