This commit adds memory allocation debugging
Searching for memory allocation faults are still difficult to find.
This commit adds the debug compilation options that makes all
memory allocations verbose. This way, a user can see exactly what
happened on what line when the daemon crashed. These messages
will only show when the user starts the daemon in debug mode and with
the DEBUG compilation flag enabled.