This commits add upnp ssdp capabilities
One feature of uPnP is ssdp. This allows programs to
poll the network for devices that match a certain
description. pilight now has the ability to respond
to those polls. This has a few (major) consequences.

1. pilight will know start by default on a random port.
You can still set the port in the settings to a fixed
number.

2. pilight will automatically determine if there is
already another pilight daemon running on the network.
If so, pilight will connect to that daemon as a node
and start cooperating with this daemon. Setting a
daemon to client-mode in the settings is thereby
deprecated. If no other daemon was found, pilight
will start itself as a daemon.

3. All pilight clients (receiver, controller and sender)
will use ssdp to automatically connect to the main
daemon. This means the -S (server) and -P (port)
arguments has been removed.

4. All GUIs should use ssdp to automatically determine
the ip and port pilight is running at. Several examples
have been given in the "clients" folder on how to do
this.