This commit greatly improves the pilight node feature.
Previously, pilight wasn't capable of handling sensors and relay's
connected to other pilight instances then the main daemon was
running on. This has changed by implementing unique UUID's for
each device pilight is running on. In the configuration file
you can define which sensor or relay is connected to which device
by defining the appropriate UUID:

{
   "location": {
      "name": "location",
      "device": {
         "name": "device",
         "uuid": "0000-00-00-00-000000",
         "protocol": [ "relay" ],
         "id": [{
            "gpio": 3
         }],
         "state": "off"
      }
   }
}

pilight nodes will now try to automatically reconnect to the
main pilight daemon when the connection is lost. This makes
using multiple raspberry pi with pilight much easier to
manage.

A new pilight client has been added called pilight-uuid.
This will output the uuid of the specific device which
can be used in the configuration file.