This commit changed the webserver authentication and user
1. This commit changes the webserver-username, webserver-password
and the webserver-authentication to a single
webserver-authentication setting. Instead of enabling the
webserver authentication by setting the webserver-authentication
setting to one, and then seperately defining the username
and password for authentication, the webserver-username and
webserver-password have been merged into the
webserver-authentication setting. This means you now enable
the webserver authentication as follows:

"webserver-authentication": [ "username", "password" ]

2. Before this commit, all PHP scripts were ran with root
privileges. Of course, this is a great security risk. This
commit changes the user the webserver runs at to www-data.
A new setting has been introduced to be able to change
this user:

"webserver-user": "www-data"