This commit makes receiving more stable
Some protocols pretty much use the same structure. However, mostly,
they are also different in a significant way. Previously, pilight
incoorporated the conflicts structure. This meant that the first
matching protocol processed the pulses and the protocols it conflict
with were dependent on this processing. The trouble began when
the first protocol processed the pulses differently from the
conflicting protocols and thereby the conflicting protocols stopped
working. This for example happened with the silvercrest / mumbi /
arctech_old protocol. pilight will now call all protocols separatly
so each of them can process the pulses independently.

To make sure the speed of receiving isn't affected by this change,
a thread safe queue has been introduced between the pulse receiving
and protocol parsing function.

The receiving thread has also got a higher CPU priority so whatever
the Raspberry Pi is doing doesn't affect pilight functionality.

On the side note. A small bug in the coords2tz function was fixed
in which the longitude and latitude where swapped.