One coding improvement and one addition
1) Made all functions and variables that should be limited
to the local scope static so we don't end up in naming
conflicts that easy.

2) Some protocols use a variable length and really large footers.
The footer filter has been increased from 34000 to 100000 and
the raw lengths can now be a range. The old rawlen is still available,
but two new protocol variables are introduces 'minrawlen' and
'maxrawlen'. When using these new variable to define a raw length
range, make sure the old 'rawlen' value contains the average of both.
These values can be used when receiving protocols that have variable
rawlengths like Ninja Blocks.