Removed Radxa support
The Radxa Rock Pro support has been removed until the platform is
more mature. At this time, it only pretty ancient kernels (3.0.y) are
(properly) supported and the SoC only native supports the Rising
and Falling interrupts. Radxa will be included again when the
a proper kernel with support for the Both interrupt is available.