Greatly improved event library
- AND and OR statements are now evaluated as one might expect. This means
that an OR can let (parts of a) rule succeed without having to parse all
logic. An AND can let (parts of a) rule fail without having to parse all
logic. This greatly increases the efficiency in rule parsing.
- All actions logic is now executing in a separate thread asynchronic.
- The switch action has been enhanced with a FOR and AFTER variable. Both
are optional. E.g.,

IF 1 == 1 THEN switch DEVICE light TO on FOR 30 AFTER 10

- The dim action has been enhanced with a FOR, AFTER, FROM, and IN variable.
For examples:

IF 1 == 1 THEN dim DEVICE wakeup TO 15 FROM 1 IN 2700
IF 1 == 1 THEN dim DEVICE dimmer TO 1 AFTER 10 FOR 30

* FOR: Change state for X seconds and then restore to previous state.
* AFTER: Don't change immediatly but after X seconds.
* FROM + IN: Change state of device FROM X to Y in Z seconds.

- If a new action is triggered for a specific device, while another action
is already running, the previous action will be aborted.

- The new event action thread API make the above easy to program for
community developers.