Allow device properties as action arguments
This commit add the possibility to use device properties
inside the action arguments. This allows you to dynamically
change the behavior of an event. A few examples:

IF 1 == 1 THEN dim DEVICE dimmer TO dimmerMax.dimlevel FOR dimmerDuration.dimlevel
IF 1 == 1 THEN switch DEVICE light TO television.state
IF 1 == 1 THEN label DEVICE dimLabel TO dimmer.dimlevel

This new feature can be used on the following fields
- AFTER
- FOR
- TO
- FROM
- IN