Added the DATE_ADD function.
This function behaves much alike the MySQL DATE_ADD function. The input
can either be a datetime device of a fixed datetime string. The format
of the datetime string should be "%Y-%m-%d %H:%M:%S". The output of
the DATE_ADD function will also be in the "%Y-%m-%d %H:%M:%S" format.
This is important if you want to use the DATE_ADD function inside
the DATE_FORMAT function. Also consider that the output will be omitted
by quotes.

The second argument contains the unit to be added or substracted from the
date. This is formatted as the unit can be: SECOND, MINUTE, HOUR
DAY, MONTH, or YEAR. The unit can be either a positive or negative
number.

A few examples (14:50:10 18-04-2015)

IF DATE_ADD(datetime, -20 YEAR) IS ...
=
IF "1995-04-18 14:50:10" IS ...

IF DATE_ADD(datetime, +10 MONTH) IS ...
=
IF "2016-02-18 14:50:10" IS ...

IF DATE_ADD(datetime, -3 DAY) IS ...
=
IF "2015-04-15 14:50:10" IS ...

IF DATE_ADD(datetime, 10 MINUTE) IS ...
=
IF "2015-04-18 15:00:10" IS ...

IF DATE_ADD(datetime, -11 SECOND) IS ...
=
IF "2015-04-18 14:49:59" IS ...

IF DATE_ADD(1995-04-18 14:50:10, +20 YEAR) IS ...
=
IF "2015-04-18 14:50:10" IS ...

This makes it easier to trigger an event based on timed events. For example,
turning on the lights 1 hour before the sunset time has been reached.

IF sunriseset.set == DATE_FORMAT(DATE_ADD(datetime, -1 HOUR), "%Y-%m-%d %H:%M:%S", %H.%M) THEN ...