This commit changes the way actions (arguments) are parsed
Let's take the following rule as an example:

IF ... THEN switch DEVICE lamp TO on FOR 10

In the previous implementation the action part was parsed by looking
at the spaces. So "DEVICE [space] lamp" meant that the parameter
DEVICE had a value "lamp". The problem with this approach was
that we could only pass single worded arguments. This was not
		possible:

IF ... THEN switch DEVICE lamp TO on FOR 10 MINUTE

The old parser would assume that MINUTE was a new parameter with
no value. One way to solve this issue was using quotes:

IF ... THEN switch DEVICE lamp TO on FOR "10 MINUTE"

But that made rule creation less fluent.

The new parser takes a different approach. Instead of looking at spaces
the new parser looks for the actual parameters of an action. In the
example, the DEVICE, TO, and FOR words are all paramters to the
action switch. Instead of looking at the spaces to split the parameters
we now look at these words. So everything between DEVICE and TO should be
a value to DEVICE. Everything between TO and FOR should be a value to TO
and everything between FOR and the end of the rule is a value to FOR.

This means we don't need quotes. The parser just looks at all values
between two parameters and assumes they are values to the previous
parameter. This does introduce a new rare problem. E.g.:

IF ... THEN sendmail MESSAGE THIS MESSAGE CAME FROM PILIGHT TO ...@... FROM ...@...

As you can see, both the MESSAGE and FROM parameters are also part of
the mail a user wanted to send. When we pass this rule to the parser
it will assume the following:

MESSAGE: THIS
MESSAGE: CAME
FROM: PILIGHT
TO: ...@...
FROM: ...@...

To make sure that certain words are parsed are values and not as parameters, we
can escape them:

IF ... THEN sendmail MESSAGE THIS \MESSAGE CAME \FROM PILIGHT TO ...@... FROM ...@...

This will parse as:
MESSAGE: THIS MESSAGE CAME FROM PILIGHT
TO: ...@...
FROM: ...@...

This new parser was made to introduced a new feature to the switch, dim,
and label actions. All FOR, AFTER, and IN parameters now use a resolution. E.g.:

IF ... THEN switch DEVICE lamp TO on AFTER 30 SECOND FOR 10 MINUTE

or with device parameters:

IF ... THEN switch DEVICE lamp TO on AFTER lampDelay.dimlevel SECOND FOR lampDuration.dimlevel MINUTE