Add infinite rule loop prevention.
Badly written rules can trigger itself in an infinite loop.
An example:

IF lamp.state IS on OR lamp.state IS off THEN toggle DEVICE lamp BETWEEN on AND off

pilight now knows when a rule triggers itself and will output an error
in these cases. This does however not prevent infinite loops like these:

IF lamp.state IS on THEN switch DEVICE lamp TO off
IF lamp.state IS off THEN switch DEVICE lamp TO on