Allow functions and device properties in action arguments
Before action arguments are passed to the actual function modules
they are first parsed by the eventing library. This allows the usage
of functions and device parameters in the action part of a rule.

A few examples:

IF 1 == 1 THEN dim DEVICE dimmer TO RANDOM(0, 10)
IF 1 == 1 THEN sendmail TO ... SUBJECT Doorbell rang @ DATE_FORMAT(datetime, %c) MESSAGE ...
IF 1 == 1 THEN switch DEVICE switch TO on FOR RANDOM(0, 10) MINUTE