Made e-mail address validation less strict
E-mail address validation in sendmail.c and settings.c was too strict.
Adresses containing other than alphabetic and numeric characters were
not accepted although they are actually valid. The regex for the
validation has been changed to verify that the e-mail address is in
the format: aaaaa@bbbbb.ccccc

The only limitations now are:

* aaaaa, bbbbb and ccccc are all mandatory and may have any length and
may contain any character except @.
* Multiple adjacent dots are not allowed. (so aaaaa@bbbb.x.ccccc is
valid, but aaaaa@bbbbb..ccccc is not).