The new base code of the pilight v7 rewrite
The goal of the new code base is incoorporating
robust software design patterns to solve common
issues involved in vastly growing and complex
software, as also happened in pilight.

Due to the impact these new patterns have in
the pilight codebase, intermediate releases
where not an option. That would have resulted
in a incomplete and non-usable pilight version.

The biggest improvements are listed in this
commit message, but there are numbers of
smaller ones to make pilight more robust.

So what changed?
- A threadpool was implemented to centralize
all threaded functionality. This means we no
longer start a thread for each invididual
function, but to be executed functions are
added to a queue that a number of workers
will execute
- A event notification library was implemented
with socket handling eventpool and a timerpool.
* The eventpool will handle all sockets and
files.
* A timerpool will handle all periodic or timed
function executions.
- An observer pattern was implemented to
decouple pilight functionality. Functions can
be executed based on various internal event
such as received socket messages, received
433.92 codes, config updates etc. The notifier
doesn't know who is listening and the listener
doesn't have to know who's listening. This also
makes pilight much more modular and more easy to
extend.

These new patterns already use each other a lot
in the new pilight codebase. Listeners are
handled by threadpool workers. Time functions
are handled by the threadpool workers after a
specific delay. Functions handled by the
threadpool can trigger new events, again
notifying new listeners that are handled by the
threadpool again. Etc.

-----------------------------------------------

- Mongoose, the previously used webserver
library was removed and the webserver was
rewriten to use the above design patterns. When
doing so, SSL was implemented as well using the
built-in mbedtls library.
- The storage backened has been made data-format
and data provider independent. This means that
additional modules can be created to let the
config be written in XML or YAML, or let the
config be stored in a database.
- UUID will be used in both sensors and relays,
as well as senders. This prevents senders to
interfere with each other in AdHoc networks.
- Multiple AdHoc networks are now allowed,
pilight will also not connect to a already
existing pilight instance already running as it
did before.

The provide the new functionality a few new
settings have been introduces.
- Each pilight instance need to be named:
"name": "NAS"
- Then we need to tell if pilight runs
as a server or client:
"adhoc-mode": "server"
"adhoc-mode": "client"
When omitting this setting, pilight will
run as a individual instance.
- In case of the client, we need to tell
to which master it needs to connect.
"adhoc-master": "NAS"
- To use the SSL webserver functionality
we need to tell pilight where to find the
SSL pem file:
"pem-file": "/etc/pilight/pilight.pem"

-----------------------------------------------

A great side-effect to the new incoorporated
patterns is the great reduction of pilight
resource consumption. Especially the CPU usage
is greatly reduced and pilight is more
responsive. This also makes pilight work better
on less powerfull devices such as routers.

-----------------------------------------------

What DOES NOT work yet!

- The new codebase hasn't been ported to Windows
yet. That will happen in the upcoming months
(hopefully with the help of some users).
- The Z-Wave module and the 433Lirc also
haven't been ported.
- The pilight logging is still using the
version 7 mechanisms.

-----------------------------------------------

A specific forum has been opened to discuss
this new pilight base at:
https://forum.pilight.org/Forum-pilight-2015-2016-rewrite