Changed pilight default license from GPLv3 to MPLv2
All code written by CurlyMo (founder of pilight) and
soley by CurlyMo is know licensed as MPLv2 instead of
GPLv3.

The mean reason for this change is that the idea of
having pieces of code decide what the license of a
whole project should be was bothering me. The MPLv2
for me finds the right balance between permissive
and copyleft licenses.

If code is licensed under GPLv3 is used, the project
as a whole has to be licenses under GPLv3.

Code released under Apache v2 can be used in closed
source project whereby any changes can be left closed
source as well.

The MPLv2 requires that code remains open source,
that all changes are also published, but allows
developers to use the code in closed source projects.

The main license of pilight, therefor doesn't change.
pilight uses code published under GPLv3 and can
therefor not be offered as a MPLv2 project. But
because MPLv2 code can be used in GPLv3 projects,
all code originally made for pilight is now licensed
under MPLv2.

Developers that has contributed to pilight before
have the freedom to change the license of their
contributed code to MPLv2 as well. For those who
want to, please do so i a pull request. Until then
all colaborative code remains licensed under GPLv3.