First commit of protocol 433 unit tests.
In the future protocols will be validated
according to all tests in this unit tests.
Whenever something changes in the protocol
code, these tests will run again to validate
proper working.

To ensure the proper working of existing
protocol, feel free to extend the unit
tests with your own pulses.