Test if socket events take place from main thread
Socket operations are not thread safe. That means
we have to make sure that anything we do on our
sockets is done from the main thread.

The socket library was (again) adapted to deal with
this requirement as well as the webserver.

This limitation is also actively monitored
with asserts and tested inside the unit tests.
Every socket test will now run from the main
thread and also from another thread.

Modules that use the pilight library will
automatically work, because of the way the
individual library components are implemented.