Add astyle linting to Travis
This adds a new configuration to Travis that runs astyle on the diff for
newly pushed commits and verifies that the specified code style is followed.
The output in the Travis build log includes suggested fixes to make the process
easier.

The aim is to eventually use astyle to lint all pilight source files, which is
is very useful for pull requests, in order to ensure a consistent codestyle.