Break up long label text into multiple lines
Currently a label text, longer than 50% of the available width, is only partially visible.

With this commit a label text is allowed to use 95% of the available width and wrap around if it is even longer.
This ensures that label text is always completely  visible, also on narrow screens.
(cherry picked from commit c3d7f460733f6382040c52f0198d7668bb5c3323)