Fix segfault in isrunning() on encountering long process name
Running the slack or spotify desktop apps while running the unittests revealed
that the process name was too long for the function to store in a char array
of size 255. Fixed by increasing the size of the buffer to 1024, and add the
bufsize as an argument to psutil_proc_name() to allocate a dynamic buffer size.