Update pilight code style guidelines
Also add git pre-commit hook script to verify that new commits adhere to the
guidelines and a shell script to verify the code style using the Artistic Style
tool and a whitelist. Currently, all files in the pilight repository are
whitelisted.

The options for AStyle are found in ci/astyle.conf; this is where the actual
coding style that the linter should use is defined.

[ci skip]