Rewritten event parser
The event library has been fully rewritten.
Were pilight previously used a non-standard
parser and interpreter design, it has now
adopted proven design patterns in programming
language design. A recursive descent parser
which generates an abstract syntax three
which is then executed by an interpreter.

This new design fixes some bugs in which
operator precedence was not parsed properly.
It also allows new features like (nested)
if/else rules. Lastly, because the parser
creates an AST on the first run, the actual
interpreting of the rule is massively faster
(about 1.5x to 2x) than in the previous event
library.

A possible downside of this new design is
that writing rules is a bit more strict in the
language rules to follow.

Additional changes are:
- All action now use the 'parsed arguments'
by default.
- Operators and functions are now allowed
to return different variable type than strings.
- New eventing tests have been created to
demonstrate the new functionality.